+++ src/actors/clipper_actor.rs
@@ -1,45 +1,42 @@
 use std::{
-    hint::unreachable_unchecked,
     path::{Path, PathBuf},
-    sync::{
-        mpsc::{Receiver, SyncSender},
-        Arc, Mutex,
-    },
+    sync::{Arc, Mutex},
 };
 
 use anyhow::Context;
+use crossbeam_channel::{Receiver, Sender};
 use log::{debug, info};
-use rayon_core::{ThreadPool, ThreadPoolBuilder};
+use once_cell::sync::OnceCell;
 
 use crate::{
     already_processed::AlreadyProcessed,
     io::{find_unused_prefix, named_tempfile, touch},
     outside::StreamTransformer,
     result::{err_msg, Result},
-    types::{Extension, Metadata, Timestamp, Timestamps},
+    types::{Extension, Timestamp},
     utils::MutexUtils,
 };
 
-use super::{Actor, DownloadedStream, VideoTitle};
+use super::{Actor, TimestampedClip, VideoTitle};
 
 #[derive(Debug)]
 pub struct ClipperActor<'a> {
-    pool: ThreadPool,
+    id: usize,
     stream_tsf: &'a dyn StreamTransformer,
     out_dir: &'a Path,
     ext: Extension,
     cache: Arc<Mutex<AlreadyProcessed>>,
 
-    receive_channel: Option<Receiver<DownloadedStream>>,
-    send_channel: Option<SyncSender<VideoTitle>>,
+    receive_channel: Option<Receiver<TimestampedClip>>,
+    send_channel: Option<Sender<VideoTitle>>,
 }
 
-impl Actor<DownloadedStream, VideoTitle> for ClipperActor<'_> {
-    fn set_receive_channel(&mut self, channel: Receiver<DownloadedStream>) {
+impl Actor<TimestampedClip, VideoTitle> for ClipperActor<'_> {
+    fn set_receive_channel(&mut self, channel: Receiver<TimestampedClip>) {
         self.receive_channel = Some(channel);
     }
 
-    fn set_send_channel(&mut self, channel: SyncSender<VideoTitle>) {
+    fn set_send_channel(&mut self, channel: Sender<VideoTitle>) {
         self.send_channel = Some(channel);
     }
 
@@ -54,94 +51,50 @@ impl Actor<DownloadedStream, VideoTitle> for ClipperActor<'_> {
             .take()
             .ok_or_else(|| err_msg("Send channel not set"))?;
 
-        debug!("Actor started, waiting for a downloaded stream");
+        debug!(
+            "{}: Actor started, waiting for a downloaded stream",
+            self.id
+        );
 
-        for DownloadedStream {
-            video_id,
-            file,
-            metadata,
-            timestamps,
+        for TimestampedClip {
+            stream_info,
+            start,
+            end,
         } in receive_channel
         {
-            debug!("Stream '{video_id}' received");
+            let video_id = &stream_info.video_id;
+            let stream_file = &stream_info.stream_file;
+            let metadata = &stream_info.metadata;
+
+            debug!("{}: Stream '{video_id}' received", self.id);
             info!(
-                "Splitting '{}' into {} clips",
+                "{}: Clipping '{}' ({} - {}) into '{}'",
+                self.id,
                 metadata.title,
-                timestamps.len()
+                start.t_start,
+                end.as_ref().map_or("END", |end| end.t_start.as_str()),
+                start.title
             );
 
-            self.create_clips(
-                file.path(),
-                self.out_dir,
-                &timestamps,
-                &metadata,
-                &video_id,
-                self.ext,
-            );
+            let out_empty = self.reserve_output_path(self.out_dir, &start.title, self.ext);
 
-            self.cache.with_lock(|mut cache| cache.push(video_id))?;
-
-            debug!("Iteration completed. Waiting for next stream");
-        }
-
-        debug!("All iterations completed. Stopping the actor.");
-        Ok(())
-    }
-}
-
-impl<'a> ClipperActor<'a> {
-    pub fn new(
-        num_threads: usize,
-        stream_tsf: &'a dyn StreamTransformer,
-        out_dir: &'a Path,
-        ext: Extension,
-        cache: Arc<Mutex<AlreadyProcessed>>,
-    ) -> Result<Self> {
-        let pool = ThreadPoolBuilder::new().num_threads(num_threads).build()?;
-        Ok(Self {
-            pool,
-            stream_tsf,
-            out_dir,
-            ext,
-            cache,
-            receive_channel: None,
-            send_channel: None,
-        })
-    }
-
-    fn create_clips(
-        &self,
-        input: &Path,
-        out_dir: &Path,
-        timestamps: &Timestamps,
-        metadata: &Metadata,
-        video_id: &str,
-        extension: Extension,
-    ) {
-        let album = format!("{} ({})", metadata.title, video_id);
-        let preprocess = |start: &Timestamp, _| {
-            let mut output = find_unused_prefix(out_dir, &start.title, extension, true)
-                .context("Could not build output file path")
-                .unwrap();
-
-            // Use the .empty extension for the placeholder
-            output.set_extension("empty");
-
-            touch(&output).unwrap();
-            output
-        };
-
-        let process = |start, end, out_empty: PathBuf| {
-            let out_tmp = named_tempfile(extension)
+            let out_tmp = named_tempfile(self.ext)
                 .context("Could not create tempfile")
                 .unwrap();
 
             // Create clip to tempfile (slow, things may go bad)
-            Self::create_clip(self.stream_tsf, input, out_tmp.path(), start, end, &album)
-                .context("Could not create clip")
-                .unwrap();
+            let album = format!("{} ({})", metadata.title, video_id);
+            self.create_clip(
+                stream_file.path(),
+                out_tmp.path(),
+                &start,
+                end.as_ref(),
+                &album,
+            )
+            .context("Could not create clip")
+            .unwrap();
 
-            let output = out_empty.with_extension(extension.with_no_dot());
+            let output = out_empty.with_extension(self.ext.with_no_dot());
 
             // When finished, move to output file (fast, nearly no errors)
             // First try to do a simple move
@@ -152,48 +105,67 @@ impl<'a> ClipperActor<'a> {
 
             // Remove the placeholder
             std::fs::remove_file(out_empty).unwrap();
-        };
 
-        self.for_each_clip(timestamps, preprocess, &process);
+            info!("{}: Clip '{}' completed", self.id, start.title);
+
+            // If last clip processed, add video_id to cache
+            if Arc::strong_count(&stream_info) == 1 {
+                self.cache
+                    .with_lock(|mut cache| cache.push(video_id.to_string()))?;
+            }
+
+            debug!("{}: Iteration completed. Waiting for next clip", self.id);
+        }
+
+        debug!("{}: All iterations completed. Stopping the actor.", self.id);
+        Ok(())
+    }
+}
+
+impl<'a> ClipperActor<'a> {
+    pub fn new(
+        id: usize,
+        stream_tsf: &'a dyn StreamTransformer,
+        out_dir: &'a Path,
+        ext: Extension,
+        cache: Arc<Mutex<AlreadyProcessed>>,
+    ) -> Result<Self> {
+        Ok(Self {
+            id,
+            stream_tsf,
+            out_dir,
+            ext,
+            cache,
+            receive_channel: None,
+            send_channel: None,
+        })
     }
 
-    /// Execture a process for every clip described by the timestamps concurrently.
+    /// Create an empty placeholder for the clip in the output directory.
     ///
-    /// This function does not create any clip itself but simply call the process function with
-    /// each starting and ending timestamps.
+    /// This will return a path to the placeholder with a ".empty" extension
+    /// such that when replacing the extension with the given one, it can be
+    /// created without worry of overwriting a file.
     ///
-    /// Before the process function is sent to run concurrently for each clip,
-    /// another preprocessing function is called in the main thread.
-    /// It is garanteed that this preprocessing function cannot be executed concurrently.
-    fn for_each_clip<'scope, T, Fpre, F>(
-        &self,
-        timestamps: &'scope Timestamps,
-        mut preprocess: Fpre,
-        process: F,
-    ) where
-        T: Send + 'scope,
-        Fpre: FnMut(&'scope Timestamp, Option<&'scope Timestamp>) -> T + Send,
-        F: Fn(&'scope Timestamp, Option<&'scope Timestamp>, T) + Sync,
-    {
-        let process = &process;
-        self.pool.scope(move |scope| {
-            for sl in timestamps.windows(2) {
-                if let [start, end] = sl {
-                    let data = preprocess(start, Some(end));
-                    scope.spawn(move |_| process(start, Some(end), data));
-                } else {
-                    // SAFETY: slice::windows(2) generates slices of 2 elements
-                    // In the future, the use of slice::array_windows would remove the test
-                    unsafe { unreachable_unchecked() }
-                }
-            }
+    /// Uses internally a lock to avoid returning the same path in two concurrent
+    /// method calls.
+    /// This however assumes that the output directory is not changing outside
+    /// of this method during the call.
+    fn reserve_output_path(&self, out_dir: &Path, title: &str, extension: Extension) -> PathBuf {
+        static LOCK: OnceCell<Mutex<()>> = OnceCell::new();
 
-            // Do not forget to do the last timestamp up to the end of the stream
-            if let Some(start) = timestamps.last() {
-                let data = preprocess(start, None);
-                scope.spawn(move |_| process(start, None, data));
-            }
-        });
+        LOCK.get_or_init(|| Mutex::new(())).with_lock(|_lock| {
+            let mut output = find_unused_prefix(out_dir, title, extension, true)
+                .context("Could not build output file path")
+                .unwrap();
+
+            // Use the .empty extension for the placeholder
+            output.set_extension("empty");
+
+            touch(&output).unwrap();
+
+            output
+        })
     }
 
     /// Create a clip of a stream.
@@ -203,7 +175,7 @@ impl<'a> ClipperActor<'a> {
     ///
     /// If `end` is not specified, clip will continue until the end of the stream.
     fn create_clip(
-        stream_tsf: &'a dyn StreamTransformer,
+        &self,
         input: &Path,
         output: &Path,
         start: &Timestamp,
@@ -214,20 +186,13 @@ impl<'a> ClipperActor<'a> {
         let out_ext = Extension::from_path(output).context("Invalid output extension")?;
         let tmp = named_tempfile(out_ext)?;
 
-        stream_tsf
+        self.stream_tsf
             .extract_clip(input, tmp.path(), start, end, album)
             .context("Could not extract a clip of the audio file from the timestamps")?;
 
-        stream_tsf
+        self.stream_tsf
             .normalize_audio(tmp.path(), output)
             .context("Could not normalize audio")?;
-
-        debug!(
-            "Clip '{}' ({} - {}) completed",
-            start.title,
-            start.t_start,
-            end.map_or("END", |end| end.t_start.as_str())
-        );
 
         Ok(())
     }