+++ src/actors/message.rs
@@ -1,6 +1,8 @@
+use std::sync::Arc;
+
 use tempfile::NamedTempFile;
 
-use crate::types::{Metadata, Timestamps};
+use crate::types::{Metadata, Timestamp, Timestamps};
 
 pub type VideoId = String;
 pub type VideoTitle = String;
@@ -11,4 +13,16 @@ pub struct DownloadedStream {
     pub file: NamedTempFile,
     pub metadata: Metadata,
     pub timestamps: Timestamps,
+}
+
+pub struct StreamInfo {
+    pub video_id: String,
+    pub stream_file: NamedTempFile,
+    pub metadata: Metadata,
+}
+
+pub struct TimestampedClip {
+    pub stream_info: Arc<StreamInfo>,
+    pub start: Timestamp,
+    pub end: Option<Timestamp>,
 }